refactor: 适配DP/HiChain新增接口+公共事件整合+代码去重
1. DP适配: 新增OnAccountAclAdd/Active回调, serviceIdList
   全链路int32→int64; OnAccountAcl*移除宏隔离与设备类型门控,
   均正常处理; DeviceAcl两路统一为onDeviceProfileDeleted;
   Add/Active改DelNotTrustDevice+onDeviceProfileAdd;
   提炼GetUdidFromProfile等helper去重
2. HiChain适配: 移除HICHAIN_USER_LEVEL_CALLBACK_ENABLE宏隔离,
   订阅正常; returnInfo解析下沉到组网层(OnGroupInactiveInUser),
   认证层只转发; 新增AuthHichainParseReturnInfo解析;
   GroupChangeListener补5字段
3. 公共事件: Login/Logout/Logoff整合为HandleDistributedAccountChange;
   清理5个仅日志事件,仅保留Switched
4. 组网层: lnn_network_manager注册onGroupInactiveInUser,
   处理同账号群组关系移除快速离线

+++ core/bus_center/service/include/bus_center_event.h
@@ -109,7 +109,7 @@ void LnnNotifyDmHookRegisteredEvent(void);
 void LnnNotifyAccountSwitchCheckEvent(int32_t userId);
 
 void LnnNotifyAccountAclChangeEvent(
-    const char *udid, int32_t localUserId, int32_t peerUserId, const int32_t *serviceIdList, uint32_t serviceIdCount);
+    const char *udid, int32_t localUserId, int32_t peerUserId, const int64_t *serviceIdList, uint32_t serviceIdCount);
 
 #ifdef __cplusplus
 }
