+++ tests/core/authentication/unittest/auth_device_profile_listener_mock.cpp
@@ -60,7 +60,7 @@ void NotifyRemoteDevOffLineByUserId(int32_t userId, const char *udid)
 }
 
 void LnnNotifyAccountAclChangeEvent(
-    const char *udid, int32_t localUserId, int32_t peerUserId, const int32_t *serviceIdList, uint32_t serviceIdCount)
+    const char *udid, int32_t localUserId, int32_t peerUserId, const int64_t *serviceIdList, uint32_t serviceIdCount)
 {
     (void)udid;
     (void)localUserId;
